Add files via upload
Embarking on my internship journey as a Machine Learning intern at Prodigy InfoTech was an exciting and educational experience. Over the course of my internship, I had the opportunity to work on a diverse range of projects that expanded my knowledge and skills in the field of machine learning.

Prodigy InfoTech: A Hub for Learning Prodigy InfoTech, at (https://prodigyinfotech.dev/), is a company that prides itself on nurturing young talent in the field of machine learning and artificial intelligence. It offers a supportive environment where interns can learn, grow, and apply their skills to solve real-world problems. My journey at Prodigy InfoTech was not just about working on projects; it was about gaining hands-on experience, mentorship, and the opportunity to contribute meaningfully to the world of AI.

1. Cats and Dogs Classification using SVM: This project involved building a Support Vector Machine (SVM) model to classify images of cats and dogs. It was a fantastic introduction to the world of image classification and SVM algorithms.

2. House Price Predictions using Linear Regression: Predicting house prices is a common machine learning task, and I had the opportunity to apply linear regression techniques to predict real estate prices.

3. Mall Customer Segmentation using K-Means: Customer segmentation is crucial for businesses, and I learned how to use K-Means clustering to segment mall customers based on their shopping behavior.

4. Hand Gesture Recognition Using Keras: The code continues with defining and training a convolutional neural network model using Keras, and then evaluates its performance. Finally, it visualizes the model's performance using matplotlib. This script is essentially a complete pipeline for image classification using a convolutional neural network. It loads the dataset, preprocesses the images, defines the model architecture, trains the model, evaluates its performance, and visualizes the results.